+++ src/app/api/v1/companies/[companyId]/zones/route.ts
@@ -4,43 +4,26 @@ import { Zones } from "@prisma/client"
 import { Params, ResponseAPI } from "@/lib/@types/types"
 
 /**
- * TODO: migrate this route to companies
- *
- * Handle the GET request to retrieve the zones related to a specific users.
+ * Handle the GET request to retrieve the zones related to a specific company
+ * in the database.
  *
  * @param {NextRequest} request - The HTTP request containing the request data.
- * @param {Params<"userId">} params - The dynamic parameter to extract the `userId`.
+ * @param {Params<"companyId">} params - The dynamic parameter to extract the `companyId`.
  * @returns {Promise<NextResponse>} - HTTP response with the zones related to the users.
  * @example
  * ```ts
- * const response = await fetch("/api/v1/users/{userId}/zonesByCompany")
+ * const response = await fetch("/api/v1/companies/{companyId}/zones")
  * const data = await response.json()
  * ```
  */
-export const GET = async (request: NextRequest, { params }: Params<"userId">): Promise<NextResponse> => {
+export const GET = async (request: NextRequest, { params }: Params<"companyId">): Promise<NextResponse> => {
     try {
-        const userId = parseInt(params.userId)
-        const company = await prisma.companies.findFirst({
-            where: {
-                Plants: {
-                    some: {
-                        UserPlants: {
-                            some: {
-                                userId: userId,
-                            },
-                        },
-                    },
-                },
-            },
-        })
+        const companyId = parseInt(params.companyId)
         const data = await prisma.zones.findMany({
             where: {
                 plant: {
-                    companyId: company?.companyId,
+                    companyId,
                 },
-            },
-            include: {
-                plant: true,
             },
         })
         return NextResponse.json<ResponseAPI<Zones[]>>({
