+++ src/app/api/v1/zones/route.ts
@@ -1,4 +1,4 @@
-import { NextResponse } from "next/server"
+import { NextRequest, NextResponse } from "next/server"
 import { prisma } from "@/lib/prisma"
 import { Zones } from "@prisma/client"
 import { ResponseAPI } from "@/lib/@types/types"
@@ -30,5 +30,50 @@ export const GET = async (): Promise<NextResponse> => {
             },
             { status: 404 }
         )
+    }
+}
+
+/**
+ * Handle the POST request to create a new zone associated with a specific plant in the database.
+ *
+ * @param {NextRequest} request - The HTTP request containing the information of the new zone.
+ * @returns {Promise<NextResponse>} - HTTP response with the newly created zone.
+ * @example
+ * ```ts
+ * const response = await fetch("/api/v1/", {
+ *   method: "POST",
+ *   body: JSON.stringify({
+ *     latitude: 40.7128,
+ *     longitude: -74.0060,
+ *     name: "New Zone",
+ *     plant: "1"
+ *   })
+ * })
+ * const data = await response.json()
+ * ```
+ */
+export const POST = async (request: NextRequest): Promise<NextResponse> => {
+    try {
+        const response = await request.json()
+        const { latitude, longitude, name, plant } = response
+        const newZone = await prisma.zones.create({
+            data: {
+                latitude,
+                longitude,
+                name,
+                plantId: parseInt(plant),
+            },
+        })
+        return NextResponse.json<ResponseAPI<Zones>>({
+            data: newZone as Zones,
+            ok: true,
+            message: "The zone was created successfuly",
+        })
+    } catch (error) {
+        return NextResponse.json<ResponseAPI<{}>>({
+            data: {},
+            ok: false,
+            message: "Failed to create new zone",
+        })
     }
 }
